project/ocr-engine/requirements.txt
# Enhanced OCR Dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# Basic OCR Dependencies
easyocr==1.7.0
pytesseract==0.3.10

# Image Processing
scikit-image==0.21.0
scipy==1.11.1

# AI and ML (Optional)
torch==2.0.1
torchvision==0.15.2

# Utilities
requests==2.31.0
urllib3==2.0.4

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Performance
numba==0.57.1
joblib==1.3.1

# Logging and Monitoring
loguru==0.7.0
psutil==5.9.5

# File Handling
pathlib2==2.3.7
watchdog==3.0.0

# Testing
pytest-cov==4.1.0
pytest-mock==3.11.1 